Added functionality to the attend button. Need to pull in from Stephen's list of whether the user is the attendee
Fixed missing commas

Events cannot be created on a new db with the error checking in create() in server/event/index.js

Added functionality to the attend button. However, cannot re-render component on the fly

debugged the view to update properly. Need to pull in from Stephen's list of whether the user is the attendee